added posts and comments

+++ server/routes/tasks.js
@@ -7,34 +7,30 @@ const router = express.Router()
 router.get('/:userId', async (req, res) => {
     try{
         // check if user exists
-        if (!(await Users.getUserById(req.params.userId))) return res.status(404).json()
-        const tasks = await Tasks.getTasks(req.params.userId)
-        res.status(200).json(tasks)
+        if (!(await Users.getUserById(req.params.userId))) return res.status(404).json('User not found')
+        console.log(req.query.id)
+        if (req.query.id === undefined){
+            const tasks = await Tasks.getTasks(req.params.userId)
+            res.status(200).json(tasks)
+        }else{
+            taskId = req.query.id
+            const task = await Tasks.getTask(req.params.userId, taskId)
+            if (!task) return res.status(404).json('Task not found')
+            res.status(200).json(task)
+        }
     }catch{
         res.status(500).json()
     }
 })
 
-router.get('/:userId/:taskId', async (req, res) => {
-    try
-    {
-        const task = await Tasks.getTask(req.params.userId, req.params.taskId)
-        if (!task) return res.status(404).json()
-        res.status(200).json(task)
-    }
-    catch
-    {
-        res.status(404).json()
-    }
-})
-
-router.post('/:userId', async (req, res) => {
+router.post('/:userId/CreateTask', async (req, res) => {
     const bodySchema = Joi.object({
         title : Joi.string().required()
     })
     const {error} = bodySchema.validate(req.body)
     if(error) return res.status(400).json(error.details[0].message)
 
+    if (!(await Users.getUserById(req.params.userId))) return res.status(404).json()
     const userId = req.params.userId;
     const title = req.body.title;
     const result = await Tasks.createTask(userId, title);
@@ -48,20 +44,20 @@ router.put('/:userId/:taskId', async (req, res) => {
         title : Joi.string().required(),
         completed : Joi.boolean().required()
     })
-    const {error} = bodySchema.validate(req.body)
+    const {error, value } = bodySchema.validate(req.body)
     // if not valid respond with Bad request
     if(error) return res.status(400).json(error.details[0].message)
-
-
-
-    task = {id : req.params.taskId, userId : req.params.userId, title : req.body.title, completed : req.body.completed }
-    const result = await Tasks.updateTask(task)
+    const {title, completed }= value
+    const { userId, taskId } = req.params
+    if (!(await Users.getUserById(userId))) return res.status(404).json()
+    const result = await Tasks.updateTask(userId,  taskId, title, completed)
     if (!result) return res.status(404).json()
-    res.status(200).json(result)
+    res.status(200).json(taskId)
 })
 
 
 router.delete('/:userId/:taskId', async (req, res) => {
+    if (!(await Users.getUserById(req.params.userId))) return res.status(404).json()
     const result = await Tasks.deleteTask(req.params.userId, req.params.taskId)
     if (!result) return res.status(404).send('task not found')
     res.status(200).json(req.params.userId)
